pyproject.toml
[project]
name = "su6i-yar"
version = "1.0.0"
description = "Su6i Yar: Smart AI Assistant for Instagram Download, Fact-Checking, and Language Learning"
readme = "README.md"
authors = [
    { name = "su6i", email = "su6i@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "yt-dlp>=2025.0.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-openai>=0.1.0",
    "httpx>=0.27.0",
    "edge-tts>=6.1.0",
    "python-dotenv>=1.0.0",
    "watchdog>=4.0.0",
    "playwright>=1.41.0",
    "instaloader>=4.11.0"
]
